Fix dumper v loader discrepancy and register.
