+++ src/ObjWriting/Game/IW3/Sound/LoadedSoundDumperIW3.h
@@ -1,11 +1,13 @@
 #pragma once
 
-#include "Asset/IAssetCreator.h"
+#include "Dumping/AbstractAssetDumper.h"
 #include "Game/IW3/IW3.h"
-#include "SearchPath/ISearchPath.h"
-#include "Utils/MemoryManager.h"
 
 namespace loaded_sound
 {
-    std::unique_ptr<AssetCreator<IW3::AssetLoadedSound>> CreateLoaderIW3(MemoryManager& memory, ISearchPath& searchPath, Zone& zone);
+    class DumperIW3 final : public AbstractAssetDumper<IW3::AssetLoadedSound>
+    {
+    protected:
+        void DumpAsset(AssetDumpingContext& context, const XAssetInfo<IW3::AssetLoadedSound::Type>& asset) override;
+    };
 } // namespace loaded_sound
